[h265] Fix GPU hang
Accord. to ITU-T H.265 7.4.2.4 "Any PPS NAL unit containing the value
of pps_pic_parameter_set_id for the active PPS RBSP for a coded
picture (and consequently for the layer containing the coded picture)
shall have the same content as that of the active PPS RBSP for the coded
picture, unless it follows the last VCL NAL unit of the coded picture
and precedes the first VCL NAL unit of another coded picture". The same
is applicable for SPS. So, we just complete current AU if we met SPS\PPS
which differ from active ones.

Signed-off-by: Dmitry Ermilov <dmitry.ermilov@intel.com>