Fix options behavior
cmake_dependent_option defines an options if and only if other option
is defined. If other option is not defined, value will be initialized
to the given default and it can _not_ be changed. As a result we lost
opportunity to use such options as ENABLE_TEXTLOG directly without
ENABLE_ALL. This change fixes the behavior. cmake_dependent_options
is not the thing to use here.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>